refactor out the list_dir function
