+++ tests/utils/test_path_utils.py
@@ -12,15 +12,8 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 
-import os
-from tempfile import NamedTemporaryFile
-
-import pytest
-from hypothesis import given
-from hypothesis.strategies import text
 
 from zenml.logger import get_logger
-from zenml.utils import path_utils
 
 logger = get_logger(__name__)
 
@@ -117,37 +110,37 @@ TEMPORARY_FILE_SEARCH_PREFIX = "a_f*.*"
 #     assert path_utils.is_gcs_path(sample_file_path) is False
 
 
-def test_list_dir_returns_a_list_of_file_names(tmp_path):
-    """list_dir should return a list of file names inside the queried directory"""
-    with NamedTemporaryFile(dir=tmp_path) as temp_file:
-        assert path_utils.list_dir(str(tmp_path)) == [temp_file.name]
+# def test_list_dir_returns_a_list_of_file_names(tmp_path):
+#     """list_dir should return a list of file names inside the queried directory"""
+#     with NamedTemporaryFile(dir=tmp_path) as temp_file:
+#         assert path_utils.list_dir(str(tmp_path)) == [temp_file.name]
 
 
-def test_list_dir_returns_a_list_of_file_paths(tmp_path):
-    """list_dir should return a list of file paths inside the queried directory"""
-    with NamedTemporaryFile(dir=tmp_path) as temp_file:
-        assert path_utils.list_dir(str(tmp_path)) == [
-            os.path.join(tmp_path, temp_file.name)
-        ]
+# def test_list_dir_returns_a_list_of_file_paths(tmp_path):
+#     """list_dir should return a list of file paths inside the queried directory"""
+#     with NamedTemporaryFile(dir=tmp_path) as temp_file:
+#         assert path_utils.list_dir(str(tmp_path)) == [
+#             os.path.join(tmp_path, temp_file.name)
+#         ]
 
 
-@pytest.fixture(scope="module")
-@given(sample_file=text())
-def test_list_dir_returns_one_result_for_one_file(tmp_path, sample_file):
-    """list_dir should return only one result, when there is only one file created"""
-    with open(sample_file, "w"):
-        assert len(path_utils.list_dir(str(tmp_path))) == 1
+# @pytest.fixture(scope="module")
+# @given(sample_file=text())
+# def test_list_dir_returns_one_result_for_one_file(tmp_path, sample_file):
+#     """list_dir should return only one result, when there is only one file created"""
+#     with open(sample_file, "w"):
+#         assert len(path_utils.list_dir(str(tmp_path))) == 1
 
 
-@pytest.fixture(scope="module")
-@given(sample_file=text())
-def test_list_dir_returns_empty_list_when_dir_doesnt_exist(
-    sample_file, tmp_path
-):
-    """list_dir should return an empty list when the directory"""
-    """doesn't exist"""
-    not_a_real_dir = os.path.join(tmp_path, sample_file)
-    assert isinstance(path_utils.list_dir(not_a_real_dir), list)
+# @pytest.fixture(scope="module")
+# @given(sample_file=text())
+# def test_list_dir_returns_empty_list_when_dir_doesnt_exist(
+#     sample_file, tmp_path
+# ):
+#     """list_dir should return an empty list when the directory"""
+#     """doesn't exist"""
+#     not_a_real_dir = os.path.join(tmp_path, sample_file)
+#     assert isinstance(path_utils.list_dir(not_a_real_dir), list)
 
 
 # @pytest.mark.xfail()
